+++ src/jloptions.h
@@ -68,6 +68,7 @@ typedef struct {
     uint64_t heap_target_increment;
     int8_t trace_compile_timing;
     int8_t trim;
+    int8_t trace_eval;
     int8_t task_metrics;
     int16_t timeout_for_safepoint_straggler_s;
     int8_t gc_sweep_always_full;